+++ ReliabilityCalc/Calculator.h
@@ -4,6 +4,6 @@ float failureRateFromMTBF(Data data);
 
 float probabilityOfWorkingWithoutFailureInTime(Data data, int time);
 
-int calculateFrPerNumber(int failures, int number);
+float calculateFrPerNumber(float failures, float number);
 
 float calculateFrPerTime(int failures, int tested, float operatingTime, float downtime);
